added 2019 Marshal Evictions data (#124)
This PR adds bbl-appended marshal evictions data for 2019. We are using data from @0xStarcat's awesome auto-geocoder for evictions data on DAP Portal, which deduplicates the data using our standard method (by exact matches of DATE + ADDRESS + UNIT + MARSHALL NAME) and geocodes using GeoSearch. The auto-geocoder was able to achieve a 97.7% match rate to BBL for all 2019 residential and commercial eviction records.

Our dream is to connect the evictions data auto-geocoder right up to nycdb so we can have a live stream of ready-to-use eviction data updated daily. However, given @0xStarcat 's schedule, this will be a dream/goal for later this year. In the meantime, users of nycdb can access the complete 2019 data here.