Fix pyproj deprecated call and reverse lat,lng order due to updated (#150)
When using the cli command I have been receiving the following deprecation warning:

```
 /usr/local/lib/python3.7/site-packages/pyproj/crs/crs.py:53: FutureWarning: '+init=<authority>:<code>' syntax is deprecated. '<authority>:<code>' is the preferred initialization method. When making the change, be mindful of axis order changes: https://pyproj4.github.io/pyproj/stable/gotchas.html#axis-order-changes-in-proj-6
```

It seems to be coming from src/nydb/geo.py:
```
wgs84 = pyproj.Proj(init="epsg:4326")
```

If I remove the 'init=', it seems to flip the order in which it returns latitude and longitude. I made the subsequent change in transform.py and fixed the one test that was referring to this call, and all the tests are passing. 

With this change and the subsequent change to flip the order of lat, lng, the deprecation warning goes away.