Allow hyphens and internal spaces back in to house numbers (#153)
So, we received a comment from an organizer at Flatbush Tenant Coalition that some tools that relied on nycdb were removing dashes from house numbers for building in Queens, and given that they were trying to send physical mail to the address, were worried about the deliverability. This PR takes out a step in the address cleaning script that removed hyphens and internal spaces from house numbers in the hpd_contacts and hpd_registrations tables.

To get a sense of the effects of this change, you can check out these tables, which show all house numbers with dashes or internal spaces that show up in these two datasets, as well as the frequency at which they occur. From my analysis, the VAST majority of these house numbers validly have the space or dash in them.