add pluto_22v1 and pluto_latest (#193)
First, this PR adds pluto_22v1 in the same we as usual. The only changes are that new columns 2020 census tract and blocks were added. They use a different (better) format than the 2010 ones used, but those original 2010 ones are unchanged. The new columns (bct2020 and bctcb2020) are zero-padded and have the 1-digit borough code instead of standard 3-digit FIPS country code.

Secondly, this PR also adds a new data for pluto_latest which pulls from this auto-updating version of pluto on open data. Right now it also is at 22v1. There are a few columns that needed to be renamed to match the direct-from-DCP versions, and it doesn't yet have the new census 2020 columns. I'll keep an eye on this to update here if there are any changes on open data.