add tax lien sale list (#242)
This PR adds the DOF Tax Lien Sale List (2019-2021) from Open Data (#241)
It's a single table that should update as new sales happen. They have 2011-2018 in PDFs, but it doesn't seem worth it to scrape that right now.
There is a column "month" that is the date the report was published formatted as MM/YYYY, I decided to use the SQL file to change it to report_date and make it a date like YYYY-MM-01.