use multi-row insert statement instead of many single-row ones (#266)
I believe that psycopg's executemany sends a separate transaction for each INSERT statement that we're passing it, and while this didn't seem to affect the speed too much when running locally, when connecting to a remote db it was extremely slow. I believe what the old psycopg2.extras.execute_values did was concatenate all of the individual insert statements then send them all at once as a single transaction, so this was much faster. With the new psycopg3 those extras aren't available, so I instead created a single INSERT statement (safely binding the parameters with mogrify) that adds multiple rows, and then this can be run with just execute. Testing locally on a full pluto table and a remote db, this appears to get us back to roughly the old speed.

To use mogrify I needed to specify the ClientCursor subclass, but that doesn't seem to be an issue.

I've now tested this change with the auto-updater and everything seems to be working and back to normal times for the jobs