add new joined complaints_and_problems file (#271)
HPD has released a new file on open data that has the complaints and complaint-problems joined. They have said that the old separate complaints and problems files will be removed in a month or so. (#269)

This PR adds this new joined file as a new table in the existing hpd_complaints nycdb dataset. I've retained the old tables for now, so that any users can have some time to transition over to the new format. I'll keep an eye on the nycdb auto-updater so that as soon as the old ones are removed from open data I can make a quick PR to just drop the old tables.

I considered trying to reconstruct the old separate tables from the new joined one, but not all the columns are available so it wouldn't be perfect, and hopefully the month notice will be enough for everyone to transition. If anyone really needs that feel free to comment here and we can try to work out a solution.