add fee charges, consolidate omo/hwo into hpd_charges (#334)
Previously we added OMO charges/invoices and HWO charges as two datasets of 1 and 2 tables. However these are all related to and there is a fourth hpd fee charges table we didn't have. So this PR consolidates everything into a single hpd_charges dataset with the 4 tables.

I was running into some problems with tests failing when dropping tables that dont exist yet, so added IF EXISTS to sql.drop_table()