add buildings & projects for hpd local law 44 (#345)
Start to address #341 by adding the building and project tables. There are a lot of table in this collection, and some annoying special characters in field names etc. that are making it a pain to add them all. So for now I'm just adding these two most important ones for a time-sensitive project, and will update later to add more.

While I was doing this I also made a couple minor additions to the create_dataset script - adding flags to the CLI to do just some parts of the process to help when adding a multi-table dataset (eg. just printing the yaml code so you can copy a single table's schema, or just adding the test data). I'll open an issue to add more of these (and do so in a cleaner way) if this seems useful to others as well.