add furman center's subsidized housing database (#351)
This PR adds the NYU Furman Center's Subsidized Housing Database. (see new wiki entry to info on the dataset)

This dataset pulls together a variety of different data sources for subsidized housing and geocodes all the locations and matches them to BBL. They have two formats for the data, but there are some encoding issues with the subsidy-level file to I've only included the BBL-level file - but that's the must useful one for the typical NYCDB analysis projects.

This is a useful dataset to have in general, but I' specifically interested in adding it now because it will be helpful in trying to determine eligibility for Good Cause Eviction protections since subsidized housing is not covered (they have their own similar protections).