Add table to dob CofO for new FOIL data (#352)
The open data portals CofO dataset only goes back to mid-2012, and JustFix recently FOILed for older data and we now have a dataset for 2000 to mid-2012 (see https://github.com/JustFixNYC/dob_certificate_occupancy for more info). This PR adds a new table for the foil data to the existing nycdb dataset. We were also missing a wiki page for documenting this dataset, so I created one and added it to the readme as well.