update boundaries files version (#368)
There was a problem that as of a couple days ago this job would fail on the nycg table. It seems that shp2psgl was unable to read it's .dbf file. Though I was able to load this fine in other programs. We were a couple versions behind anyway, so I updated 24b to 24d and all of those files are working fine.

I updated the test data, and added zip to the installs for docker so that the boundaries test data generation script works within the container.