update dataset url for new nyc open data domain (#369)
Yesterday we got an error on download the dob_complaints dataset - it is a certificate error, and after reporting it via NYC Open Data portal the DOB person explained that the nycopendata.socrata.com domain is old and we should update to use data.cityofnewyork.us instead. And this seems to work. (Though I also got an error at first when trying to download it directly from the portal using the export button, so not totally sure and may need to revisit this).

dob_complaints and dob_certificate_occupancy were the only two datasets still using this old domain.