Add DOB NOW table to existing certificate of occupancy dataset (#371)
There is an additional table of CofO data that's managed under the newer DOB NOW system. This PR adds it as an additional table in the existing dob_certificate_occupancy dataset.

Also, in the github action workflow for publishing to PyPi there is a package (upload-artifacts) that required a version update