add a non-clipped to shoreline version of boundaries (#374)
This adds a two new datasets just like the existing boundaries one, but using the version of the files that includes water areas (ie/ not clipped to the shoreline). There are a few tables that DCP doesn't provide "water included" versions of, but otherwise everything is the same except the tables have the wi suffix, following DCPs naming conventions. The two versions of the dataset are just like we have for pluto_latest and pluto_24v1. boundaries_water_included_25a is the static version for 25a edition, and boundaries_water_included can be freely updated to the latest version. Unlike pluto, the urls need to be manually updated to bump the version though, so we can just do this whenever needed. Most versions I don't think make any meaningful changes, but eventually with the new census and political redistricting there will be major changes so it will be helpful to have a static version as well.

This also adds a static edition version of the clipped boundaries file as well: boundaries_25a