add archived copy of zipcode boundaries (#384)
Adds an archived copy of nyc zip code boundaries. This dataset was originally published on NYC Open Data Portal, but has since been removed. BetaNYC shared an archived version from when they built their boundaries map, and I've put it on JustFix's S3 so we can use it here. I searched around and was unable to find a good alternative to this, and even though it's not current or being updated I think it might still be useful in some cases. At JustFix we want to use it for a new feature on Who Owns What as a building block for defining custom areas on a map. I've added a wiki page for the dataset, and explain that it's an archived version, and link to metadata that is still available from the City's nyc-geo-metadata repo