add SQL-only dataset with dependencies for pluto_latest_districts (#386)
This adds two new datasets that are solely based on SQL scripts and other nycdb datasets, with no source data downloaded. This is a new pattern, and required some small changes to skip standard processes like creating the empty table before import. It also adds a new section to the dataset yml files to list other nycdb datasets as dependencies, which are checked before the "import" step runs.

The tables that are created via SQL are still listed in the yml schema so that the --drop and --verify commands still works, there just is no fields section for those schemas.